requirements.txt
streamlit
face-recognition
opencv-python
numpy
bcrypt
python-dotenv
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0